+++ recipes/write-root-my-dot-cnf.rb
@@ -10,8 +10,11 @@ host = db_info[:host]
 username = db_info[:username]
 password = db_info[:password]
 port = db_info[:port]
-database_name = db_info[:database]
 
+# stack deployment config includes a read-only db endpoint value
+# for use by certain mysql commands, e.g. mysqldump
+# fall back to standard endpoint if not present in stack config
+readonly_op_host = db_info[:readonly_op_host] || db_info[:host]
 
 template %Q|/root/.my.cnf| do
   source 'my.cnf.erb'
@@ -20,6 +23,7 @@ template %Q|/root/.my.cnf| do
   mode '0600'
   variables({
     host: host,
+    readonly_op_host: readonly_op_host,
     username: username,
     password: password,
     is_db_node: is_db_node,